requirements.txt
aiohappyeyeballs==2.4.3
aiohttp
aiosignal==1.3.1
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
netifaces==0.10.6
propcache==0.2.0
python-engineio==4.10.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.2.3
wsproto==1.2.0
yarl==1.18.0
